Refactor: Modularize API routes in api.js
I've separated the Elysia API routes into dedicated modules within a new `routes` directory.

- I created `routes/navigation.js` for navigation-related endpoints.
- I created `routes/mediaKeys.js` for media key simulation endpoints.
- I created `routes/shortlinksAdmin.js` for managing shortlinks.

The main `api.js` file has been updated to import and use these modules. Dependencies (`mainWindow` and `store`) are passed to the route modules as function arguments.

This change improves code organization and maintainability of the API layer.
