+++ routes/.gitkeep
@@ -1,0 +1,2 @@
+# This file is intentionally left blank.
+# It is used to ensure that the 'routes' directory is tracked by Git.
